+++ CHANGELOG.rst
@@ -4,6 +4,7 @@ Change Log
 
 Unreleased
 ----------
+* Add `filter_lfaa_by_distance()` and `get_neighbour_lfaa()` to `LowStation`.
 
 0.0.1
 -----